refactor(agents): simplify native backend resolution
- Drop parallel _SDK_PROBE map; discover provider SDK via _import_* helpers
- Move supports_native onto LLMClient; module wrapper delegates to client
- Unify _resolve_provider_type with _get_provider lookup path
- Merge generator/embedding resolve modules into giskard.agents.resolve
