+++ libs/giskard-agents/src/giskard/agents/embeddings/__init__.py
@@ -1,7 +1,6 @@
 from .base import BaseEmbeddingModel
 from .litellm_embedding_model import LitellmEmbeddingModel
 from .litellm_package_embedding_model import LiteLLMEmbeddingModel
-from .resolve import resolve_embedding_model
 
 # Default embedding model uses giskard-llm (via LitellmEmbeddingModel)
 EmbeddingModel = LitellmEmbeddingModel
@@ -11,5 +10,4 @@ __all__ = [
     "LitellmEmbeddingModel",
     "LiteLLMEmbeddingModel",
     "EmbeddingModel",
-    "resolve_embedding_model",
 ]
